+++ app/controllers/oauth/zoho_controller.rb
@@ -58,7 +58,8 @@ class Oauth::ZohoController < ApplicationController
         provider: oauth_provider,
         uid: identity[:account_id],
         email: identity[:email],
-        name: identity[:name]
+        name: identity[:name],
+        allow_create: sign_in_allow_create?
       )
     )
   end